refactor: rename 'store/user/types.ts' to 'store/user/action-types.ts'
- change type ID with 'Types.ID'.
