+++ src/store/user/action-types.ts
@@ -1,0 +1,8 @@
+import IUser from 'Entities/user-entities';
+
+export interface IUserState {
+  user: IUser;
+  loggingIn: boolean;
+  alertMessage?: string;
+  errorMessage?: string;
+}
