--- src/store/user/types.ts
@@ -1,8 +1,0 @@
-import { IUser } from 'entities/user-entities';
-
-export interface IUserState {
-  user: IUser;
-  loggingIn: boolean;
-  alertMessage?: string | null;
-  errorMessage?: string | null;
-}
